Migrations galore!
Migration to podman-compose is finished with this patch. Likewise,
cleaned up the mess that was the Makefile, django properly migrates all
databases and creates tables appropriately, and podman-compose
environment variables files are now contained in respective directories
on a per-environment basis (currently has only .env.dev but overtime
will include staging and prod).